# perf: update to zustand 4 to prevent re-renders