--- a/frontend/package.json
+++ b/frontend/package.json
@@ -73,7 +73,7 @@
     "uuid": "^8.3.2",
     "validator": "^13.7.0",
     "web-vitals": "^2.1.2",
-    "zustand": "^3.6.5"
+    "zustand": "^4.1.1"
   },
   "scriptComments": {
     "clean:emotion-types": [